separated eslint message filter spec into several files by the filtering option
